TGP-3172: move fie
